refactor ProductDetailGallery
- remove useResizeWidthEffect hook
